requirements.txt
beautifulsoup4==4.9.3
python-dotenv==0.19.0
soupsieve==2.2.1
